+++ src/components/shared/select/SelectOptionSelected.tsx
@@ -7,11 +7,13 @@ import { useSelectContext } from '@/contexts/select'
 interface SelectOptionSelectedProps {
   placeholder?: string
   className?: string
+  value?: string
 }
 
 export function SelectOptionSelected({
   placeholder,
   className,
+  value,
 }: SelectOptionSelectedProps) {
   const { itemSelected } = useSelectContext()
 
@@ -19,12 +21,14 @@ export function SelectOptionSelected({
     <div
       className={twMerge(
         'text-sm tracking-[0.02625rem] text-gray-300 overflow-hidden ',
-        itemSelected && 'text-white',
+        (itemSelected || value) && 'text-white',
         className
       )}
     >
       <p className="overflow-hidden text-ellipsis whitespace-nowrap">
-        {itemSelected ? itemSelected.label : placeholder}
+        {value && value}
+        {!value && itemSelected && itemSelected.label}
+        {!value && !itemSelected && placeholder}
       </p>
     </div>
   )